backend/requirements.docker.txt
asgiref==3.6.0
Django==4.2
psycopg2==2.9.3
python-decouple==3.8
sqlparse==0.4.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==3.14.0